Muzkanal anons add

+++ app/Acme/Http/Admin/Views/schedule/index.blade.php
@@ -1,5 +1,5 @@
 @extends('Admin::layouts.default')
-@section('title', "Schedule")
+@section('title', "Программа передач")
 
 @section('content')
 
@@ -16,75 +16,30 @@
         </div>
         <div class="x_content program">
 
-            @foreach($schedules as $key=>$schedule)
-              <div class="@if($key%13==0) first @endif program-block">
-                <div class="program-date">
-                  <h4>{{ $schedule->getMonthFull() }}</h4>
-                  <h1>{{ $schedule->getDay() }}</h1>
-                  <h5>{{ $schedule->getWeekRuFull() }}</h5>
-                </div>
-                <ul class="actions">
-                  <li><a class="btn btn-default action-view" href="{{ route('admin.schedule.show', $schedule) }}"><i class="fa fa-eye"></i></a></li>
-                  <li><a class="btn btn-default action-edit" href="{{ route('admin.schedule.edit', $schedule) }}"><i class="fa fa-edit"></i></a></li>
-                  <li>
-                    {!! Form::open(['route' => ['admin.schedule.destroy', $schedule], 'method' => 'DELETE', 'onsubmit' => "return confirm('Вы уверены ?')"]) !!}
-                    <button type="submit" class="btn btn-default action-delete" href="#" style="margin: 0;padding: 5px 8px;font-size: 18px;">
-                      {{--<span class="glyphicon glyphicon-trash"></span>--}}
-                      <i class="fa fa-trash"></i>
-                    </button>
-                    {!! Form::close() !!}
-                  </li>
-                </ul>
-              </div>
-            @endforeach
-
-
-
-
-
-          <ul id="tabs" class="nav nav-tabs teletabs hidden" data-tabs="tabs">
-            @foreach($programs as $key => $program)
-              @foreach($schedules as $schedule)
-                @if($program['date'] == $schedule->date)
-                  <li class="@if($key == 0) active @endif"><a href="#{{ $program['date'] }}" data-toggle="tab">{{ $schedule->getDay().' '.$schedule->getMonth().' - '.$schedule->getWeekRu() }}</a></li>
-                @endif
-              @endforeach
-            @endforeach
-          </ul>
-
-          <div id="my-tab-content" class="tab-content hidden">
-
-            @foreach($programs as $key => $program)
-
-              @foreach($schedules as $schedule)
-
-                @if($program['date'] == $schedule->date)
-
-                  <div class="tab-pane @if($key == 0) active @endif" id="{{ $schedule->date }}">
-
-                    <table class="table table-bordered table-striped program">
-                      <tbody>
-                        @for($i=0; $i<count($program)-1; $i++)
-                          <tr class="tele-row">
-                            <th class="tele-time">{{ $program[$i]->time }}</th>
-                            <td class="tele-show">
-                              <h4>{{ $program[$i]->name }}</h4>
-                              {{--<h5 class="tele-extra"><i class="fa fa-play-circle-o"></i>Сериал</h5>--}}
-                            </td>
-                          </tr>
-                        @endfor
-                      </tbody>
-                    </table>
-
+            @if($schedules)
+              @foreach($schedules as $key=>$schedule)
+                <div class="@if($key%13==0) first @endif program-block">
+                  <div class="program-date">
+                    <h4>{{ $schedule->getMonthFull() }}</h4>
+                    <h1>{{ $schedule->getDay() }}</h1>
+                    <h5>{{ $schedule->getWeekRuFull() }}</h5>
+                    <a href="#">@if($schedule->channel_id) {{ $schedule->channel()->first()->display }} @endif</a>
                   </div>
-
-                @endif
-
+                  <ul class="actions">
+                    <li><a class="btn btn-default action-view" href="{{ route('admin.schedule.show', $schedule) }}"><i class="fa fa-eye"></i></a></li>
+                    <li><a class="btn btn-default action-edit" href="{{ route('admin.schedule.edit', $schedule) }}"><i class="fa fa-edit"></i></a></li>
+                    <li>
+                      {!! Form::open(['route' => ['admin.schedule.destroy', $schedule], 'method' => 'DELETE', 'onsubmit' => "return confirm('Вы уверены ?')"]) !!}
+                      <button type="submit" class="btn btn-default action-delete" href="#" style="margin: 0;padding: 5px 8px;font-size: 18px;">
+                        {{--<span class="glyphicon glyphicon-trash"></span>--}}
+                        <i class="fa fa-trash"></i>
+                      </button>
+                      {!! Form::close() !!}
+                    </li>
+                  </ul>
+                </div>
               @endforeach
-
-            @endforeach
-
-          </div>
+            @endif
 
           </div>
         </div>
